css/css-values/integer_interpolation_round_half_towards_positive_infinity_order.html

<!doctype html>
<title>Testing if integer interpolation is rounded towards positive infinity</title>
<link rel="author" title="Joonghun Park" href="pjh0718@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-values-4/#combine-integers">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>

#flex-container {
    display: flex;
    animation: anim-order 4s steps(4) forwards 1;
    animation-delay: -1s;
    animation-play-state: paused;
}

@property --o {
  syntax: '<number>';
  initial-value: 0;
  inherits: true;
}

@keyframes anim-order {
  from {
    order: -2;
    --o: -2;
  }

  to {
    order: 0;
    --o: 0;
  }
}

</style>
<div id="flex-container"></div>
<script>
var test_description = "Integer interpolation should be rounded towards positive infinity";
test(
    t => {
        const container = document.getElementById("flex-container");
        const o_value = Number.parseFloat(getComputedStyle(container).getPropertyValue('--o'));
        const order_value = Number.parseFloat(getComputedStyle(container).getPropertyValue('order'));

        const rounded_o_value = Math.round(o_value);
        assert_equals(order_value, rounded_o_value, "Interpolation result for order should be rounded towards positive infinity");
    },
    test_description
);
</script>
